+++ app/models/AccountStats.py
@@ -4,10 +4,10 @@ class AccountStats(models.Model):
     'Account',
     on_delete=models.DO_NOTHING
   )
-  statuses_count = models.IntegerField(default=0)
-  following_count = models.IntegerField(default=0)
-  followers_count = models.IntegerField(default=0)
-  lock_version = models.IntegerField
+  statuses_count = models.PositiveBigIntegerField(default=0)
+  following_count = models.PositiveBigIntegerField(default=0)
+  followers_count = models.PositiveBigIntegerField(default=0)
+  lock_version = models.PositiveBigIntegerField
   last_status_at = models.DateTimeField(blank=True)
   created_at = models.DateTimeField(auto_now_add=True)
   updated_at = models.DateTimeField(auto_now=True)
