+++ app/models/AccountTagStats.py
@@ -4,7 +4,7 @@ class AccountTagStats(models.Model):
     'Tag',
     on_delete=models.DO_NOTHING
   )
-  accounts_count = models.IntegerField(default=0)
+  accounts_count = models.PositiveBigIntegerField(default=0)
   hidden = models.BooleanField(default=False)
   created_at = models.DateTimeField(auto_now_add=True)
   updated_at = models.DateTimeField(auto_now=True)
