+++ app/models/PreviewCard.py
@@ -3,14 +3,14 @@ class PreviewCard(models.Model):
   url = models.URLField(unique=True)
   title = models.TextField
   description = models.TextField
-  type = models.IntegerField
+  type = models.PositiveBigIntegerField
   html = models.TextField
   author_name = models.TextField
   author_url = models.URLField
   provider_name = models.TextField
   provider_url = models.URLField
-  width = models.IntegerField
-  height = models.IntegerField(blank=True)
+  width = models.PositiveBigIntegerField
+  height = models.PositiveBigIntegerField(blank=True)
   embed_url = models.URLField
   image_file_name = models.TextField(blank=True)
   image_content_type = models.TextField(blank=True)
