+++ app/models/Relay.py
@@ -1,7 +1,7 @@
 from django.db import models
 class Relay(models.Model):
   inbox_url = models.URLField
-  state = models.IntegerField
+  state = models.PositiveBigIntegerField
   follow_activity_id = models.TextField(blank=True)
   created_at = models.DateTimeField(auto_now_add=True)
   updated_at = models.DateTimeField(auto_now=True)
